+++ index.html
@@ -58,8 +58,7 @@
         <div class="wall-right"></div>
     </div>
 
-    <script src="logs.js"></script>
-    <script src="main.js"></script>
+    <script src="main.js" type="module"></script>
 
 </body>
 </html>
